fix casting

+++ examples/exampleAppSource/example_app.cpp
@@ -295,7 +295,7 @@ void app::update(PedalExampleApp* app) {
     ImGui::End();
 
     ImGui::SetNextWindowPos(ImVec2{160.0f,0.0f});
-    ImGui::SetNextWindowSize(ImVec2{window_w - 160, float(window_h)});
+    ImGui::SetNextWindowSize(ImVec2{static_cast<float>(window_w - 160), static_cast<float>(window_h)});
     ImGui::Begin("Right Window", nullptr, flags);
     for (int i = 0; i < NUM_SLIDERS_MAX; i += 1) {
         slider* s = app->sliders + i;
